+++ src/perl/lib/WTSI/NPG/Genotyping/Fluidigm/Subscriber.pm
@@ -3,7 +3,7 @@ package WTSI::NPG::Genotyping::Fluidigm::Subscriber;
 
 use Moose;
 use JSON;
-use List::AllUtils qw(all natatime reduce uniq);
+use List::AllUtils qw(all reduce uniq);
 use Try::Tiny;
 
 use WTSI::NPG::Genotyping::Call;
@@ -12,6 +12,7 @@ use WTSI::NPG::Genotyping::Fluidigm::AssayResultSet;
 use WTSI::NPG::Genotyping::SNP;
 use WTSI::NPG::Genotyping::SNPSet;
 use WTSI::NPG::iRODS;
+use WTSI::NPG::iRODS::Metadata; # has attribute name constants
 
 our $VERSION = '';
 
@@ -20,12 +21,7 @@ our $NO_CALL_GENOTYPE = 'NN';
 
 our $CHROMOSOME_JSON_ATTR = 'chromosome_json';
 
-# The largest number of bind variables iRODS supports for 'IN'
-# queries.
-our $BATCH_QUERY_CHUNK_SIZE = 100;
-
-with 'WTSI::DNAP::Utilities::Loggable', 'WTSI::NPG::Annotation',
-  'WTSI::NPG::Genotyping::Annotation', 'WTSI::NPG::Genotyping::Subscription';
+with 'WTSI::DNAP::Utilities::Loggable', 'WTSI::NPG::Genotyping::Subscription';
 
 has '_plex_name_attr' =>
   (is            => 'ro',
@@ -33,125 +29,48 @@ has '_plex_name_attr' =>
    init_arg      => undef,
    default       => sub {
        my ($self) = @_;
-       return $self->fluidigm_plex_name_attr;
+       return $FLUIDIGM_PLEX_NAME;
    },
    lazy          => 1,
    documentation => 'iRODS attribute for QC plex name');
 
-=head2 get_assay_resultset
 
-  Arg [1] : Str sample identifier (dcterms:identifier)
-  Arg [n] : Optional additional query specs as ArrayRefs.
-  Example : $sub->get_assay_resultset('qc', '0123456789',
-                                      [study => 12345]);
-  Description: Fetch an assay result by SNP set, sample and other optional
-               criteria. Raises an error if the query finds >1 result set.
-  Returntype : WTSI::NPG::Genotyping::Fluidigm::AssayResultSet
-
-=cut
-
-sub get_assay_resultset {
-    my ($self, $sample_identifier, @query_specs) = @_;
-    my $resultsets = $self->get_assay_resultsets
-      ([$sample_identifier], @query_specs);
-
-    my $num_samples = scalar keys %$resultsets;
-    if ($num_samples > 1) {
-      $self->logconfess("The assay results query returned data for >1 ",
-                        "sample: [", join(q{, }, keys %$resultsets, "]"));
-    }
-
-    my @resultsets = @{$resultsets->{$sample_identifier}};
-    my $num_resultsets = scalar @resultsets;
-    if ($num_resultsets > 1) {
-      $self->logconfess("The assay results query was not specific enough; ",
-                        "$num_resultsets result sets were returned: [",
-                        join(q{, }, map { $_->str } @resultsets), "]");
-    }
-
-    return shift @resultsets;
-}
-
-=head2 get_assay_resultsets
+=head2 get_assay_resultsets_and_vcf_metadata
 
   Arg [1]    : ArrayRef[Str] sample identifier (dcterms:identifier)
   Arg [n]    : Optional additional query specs as ArrayRefs.
 
   Example    : $sub->get_assay_resultsets(['0123456789'], [study => 12345]);
   Description: Fetch assay result sets by SNP set, sample and other optional
-               criteria.
-  Returntype : HashRef[ArrayRef[
-                 WTSI::NPG::Genotyping::Fluidigm::AssayResultSet]] indexed
+               criteria. Also finds associated VCF metadata.
+  Returntype : - HashRef[ArrayRef[
+                  WTSI::NPG::Genotyping::Fluidigm::AssayResultSet]] indexed
                by sample identifier. Each ArrayRef will usually contain one
                item, but may contain more if multiple assays match the
                search criteria.
+               - HashRef[ArrayRef[Str]] containing VCF metadata.
 
 =cut
 
-sub get_assay_resultsets {
+sub get_assay_resultsets_and_vcf_metadata {
   my ($self, $sample_identifiers, @query_specs) = @_;
 
-  defined $sample_identifiers or
-    $self->logconfess('A defined sample_identifiers argument is required');
-  ref $sample_identifiers eq 'ARRAY' or
-    $self->logconfess('The sample_identifiers argument must be an ArrayRef');
-  _are_unique($sample_identifiers) or
-    $self->logconfess('The sample_identifiers argument contained duplicate ',
-                      'values: [', join(q{, }, @$sample_identifiers), ']');
+  my @obj_paths =
+      $self->find_object_paths($sample_identifiers, @query_specs);
 
-  my $num_samples = scalar @$sample_identifiers;
-  my $chunk_size = $BATCH_QUERY_CHUNK_SIZE;
-
-  $self->debug("Getting results for $num_samples samples in chunks of ",
-               $chunk_size);
-
-  my @obj_paths;
-  my $iter = natatime $chunk_size, @$sample_identifiers;
-  while (my @ids = $iter->()) {
-    my @id_obj_paths = $self->irods->find_objects_by_meta
-      ($self->data_path,
-       [$self->_plex_name_attr => $self->snpset_name],
-       [$self->dcterms_identifier_attr => \@ids, 'in'], @query_specs);
-    push @obj_paths, @id_obj_paths;
-  }
-
-  my @resultsets = map {
-    WTSI::NPG::Genotyping::Fluidigm::AssayResultSet->new
-        (WTSI::NPG::Genotyping::Fluidigm::AssayDataObject->new
-         ($self->irods, $_));
+  # generate array of AssayDataObjects
+  # use to construct indexed AssayResultSets *and* find metadata
+  my @data_objects = map {
+      WTSI::NPG::Genotyping::Fluidigm::AssayDataObject->new
+            ($self->irods, $_);
   } @obj_paths;
-
-  # Index the results by sample identifier. The identifier is
-  # guaranteed to be present in the metadata because it was in the
-  # search criteria. No assumptions can be made about the order in
-  # which iRODS has returned the results, with respect to the
-  # arguments of the 'IN' clause.
-  my %resultsets_index;
-  foreach my $sample_identifier (@$sample_identifiers) {
-    unless (exists $resultsets_index{$sample_identifier}) {
-      $resultsets_index{$sample_identifier} = [];
-    }
-
-    my @sample_resultsets = grep {
-      $_->data_object->get_avu($self->dcterms_identifier_attr,
-                               $sample_identifier) } @resultsets;
-
-    $self->debug("Found ", scalar @sample_resultsets, " resultsets for ",
-                 "sample '$sample_identifier'");
-
-    # Sanity check that the contents are consistent
-    my @sample_names = map { $_->sample_name } @sample_resultsets;
-    unless (all { $_ eq $sample_names[0] } @sample_names) {
-      $self->logconfess("The resultsets found for sample AVU value ",
-                        "'$sample_identifier': did not all have the same ",
-                        "sample name in their data: [",
-                        join(q{, }, @sample_names), "]");
-    }
-
-    push @{$resultsets_index{$sample_identifier}}, @sample_resultsets;
-  }
-
-  return \%resultsets_index;
+  my @resultsets = map {
+      WTSI::NPG::Genotyping::Fluidigm::AssayResultSet->new($_);
+  } @data_objects;
+  my $resultsets_index =
+      $self->find_resultsets_index(\@resultsets, $sample_identifiers);
+  my $vcf_meta = $self->vcf_metadata_from_irods(\@data_objects);
+  return ($resultsets_index, $vcf_meta);
 }
 
 =head2 get_calls
@@ -185,7 +104,7 @@ sub get_calls {
   my @resultsets = @{$assay_resultsets};
   my $num_resultsets = scalar @resultsets;
 
-  my @sample_names = map { $_->sample_name } @resultsets;
+  my @sample_names = map { $_->canonical_sample_id } @resultsets;
   unless (all { $_ eq $sample_names[0] } @sample_names) {
     $self->logconfess('The assay_resultsets must belong to the same sample: [',
                       join(q{, }, @sample_names, ']'));
@@ -237,6 +156,22 @@ sub get_calls {
   return \@calls;
 }
 
+
+=head2 platform_name
+
+  Arg [1] : None
+  Example : my $name = $sub->platform_name();
+  Description: Return an identifier string for the genotyping platform;
+               in this case, 'fluidigm'. Used to construct a default
+               callset name in the Subscription role.
+  Returntype : Str
+
+=cut
+
+sub platform_name {
+    return 'fluidigm';
+}
+
 sub _build_calls_at {
   my ($self, $assay_address, $resultsets) = @_;
 
@@ -252,7 +187,7 @@ sub _build_calls_at {
     else {
       $self->trace("Adding fluidigm result at '$assay_address'");
 
-      my $snp = $self->_snpset->named_snp($result->snp_assayed);
+      my $snp = $self->read_snpset->named_snp($result->snp_assayed);
       push @calls,
         WTSI::NPG::Genotyping::Call->new(snp      => $snp,
                                          genotype => $result->canonical_call,
@@ -296,11 +231,11 @@ specific samples from iRODS.
 
 =head1 AUTHOR
 
-Keith James <kdj@sanger.ac.uk>
+Keith James <kdj@sanger.ac.uk>, Iain Bancarz <ib5@sanger.ac.uk>
 
 =head1 COPYRIGHT AND DISCLAIMER
 
-Copyright (C) 2014, 2015 Genome Research Limited. All Rights Reserved.
+Copyright (C) 2014, 2015, 2016 Genome Research Limited. All Rights Reserved.
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the Perl Artistic License or the GNU General