+++ src/perl/lib/WTSI/NPG/Genotyping/Sequenom/AssayResultSet.pm
@@ -6,10 +6,10 @@ use List::AllUtils qw(uniq);
 use Moose;
 use Text::CSV;
 
+use WTSI::NPG::iRODS::Metadata; # has attribute name constants
 use WTSI::NPG::Genotyping::Sequenom::AssayResult;
 
-with 'WTSI::DNAP::Utilities::Loggable', 'WTSI::NPG::iRODS::Storable',
-  'WTSI::NPG::Genotyping::Annotation';
+with 'WTSI::DNAP::Utilities::Loggable', 'WTSI::NPG::iRODS::Storable';
 
 our $VERSION = '';
 
@@ -75,7 +75,6 @@ sub canonical_sample_id {
   return shift @names;
 }
 
-
 sub snpset_name {
   my ($self) = @_;
 
@@ -84,7 +83,7 @@ sub snpset_name {
                       "' is not in iRODS");
 
   my @snpset_names = $self->data_object->find_in_metadata
-    ($self->sequenom_plex_name_attr);
+    ($SEQUENOM_PLEX_NAME);
   my $num_names = scalar @snpset_names;
 
   $num_names > 0 or
@@ -142,7 +141,31 @@ sub _build_assay_results {
 
   close $fh or $self->logwarn("Failed to close a string handle");
 
+  $records = $self->_remove_redundant_assay_results($records);
+
   return $records;
+}
+
+sub _remove_redundant_assay_results {
+    # discard any AssayResults which differ only by the 'allele' field
+    # The 'allele' field:
+    # - records the reference allele
+    # - is not used in subsequent creation of Call objects
+    # - is redundant given the assay_id and snpset manifest
+    my ($self, $raw_results) = @_;
+    my @raw_results = @{$raw_results};
+    my @unique_results;
+    foreach my $ar_i (@raw_results) {
+        my $unique = 1;
+        foreach my $ar_j (@unique_results) {
+            if ($ar_i->equivalent_within_allele($ar_j)) {
+                $unique = 0;
+                last;
+            }
+        }
+        if ($unique) { push @unique_results, $ar_i; }
+    }
+    return \@unique_results;
 }
 
 sub _parse_assay_results {