+++ src/perl/t/WTSI/NPG/Genotyping/Fluidigm/AssayResultSetTest.pm
@@ -6,7 +6,7 @@ package WTSI::NPG::Genotyping::Fluidigm::AssayResultSetTest;
 use strict;
 use warnings;
 
-use base qw(Test::Class);
+use base qw(WTSI::NPG::Test);
 use File::Spec;
 use Test::More tests => 68;
 use Test::Exception;
@@ -118,8 +118,8 @@ sub assay_addresses : Test(1) {
              'Contains expected assay addresses') or diag explain \@addresses;
 }
 
-sub sample_name : Test(1) {
-  is($resultset->sample_name, 'ABC0123456789', 'Correct sample name');
+sub canonical_sample_id : Test(1) {
+  is($resultset->canonical_sample_id, 'ABC0123456789', 'Correct sample ID');
 }
 
 sub snp_names : Test(1) {