+++ src/perl/t/identity_check_sample_wip_bayesian.t
@@ -1,0 +1,9 @@
+
+use utf8;
+
+use strict;
+use warnings;
+
+use WTSI::NPG::Genotyping::QC_wip::Check::SampleIdentityBayesianTest;
+
+WTSI::NPG::Genotyping::QC_wip::Check::SampleIdentityBayesianTest->runtests;